// src/type.d.ts
/*
 * @Author: Lanrri
 * @Email: lanrri@163.com
 * @Date: 2023-03-16 15:49:48
 * @Description: 
 * @LastEditors: Lanrri
 * @LastEditTime: 2023-03-16 15:49:55
 */

export interface CustomType {
  readonly emoji: string;
  readonly type: string;
  readonly name: string;
  readonly description: string;
}